Moves `LocalImageLoader` and `ImageDataStore` to production
