+++ FeedModules/Core Layer/FeedCoreModule/FeedCache/ImageDataStore.swift
@@ -1,0 +1,14 @@
+//
+//  ImageDataStore.swift
+//  FeedCoreModule
+//
+//  Created by Omran Khoja on 3/10/22.
+//
+
+import Foundation
+
+public protocol ImageDataStore {
+    typealias Result = Swift.Result<Data?, Error>
+    
+    func retrieve(dataForURL url: URL, completion: @escaping (Result) -> Void)
+}
