+++ FeedModules/Core Layer/Tests/FeedCoreModuleTests/Feed Cache/LocalImageLoaderTests.swift
@@ -8,58 +8,6 @@
 import XCTest
 import FeedCoreModule
 
-protocol ImageDataStore {
-    typealias Result = Swift.Result<Data?, Error>
-    
-    func retrieve(dataForURL url: URL, completion: @escaping (Result) -> Void)
-}
-
-final class LocalImageLoader {
-    private final class Task: ImageLoaderTask {
-        private var completion: ((ImageLoader.Result) -> Void)?
-        
-        init(_ completion: @escaping (ImageLoader.Result) -> Void) {
-            self.completion = completion
-        }
-        
-        func complete(with result: ImageLoader.Result) {
-            completion?(result)
-        }
-        
-        func cancel() {
-            preventFurtherCompletions()
-        }
-        
-        private func preventFurtherCompletions() {
-            completion = nil
-        }
-    }
-    
-    public enum Error: Swift.Error {
-        case failed
-        case notFound
-    }
-    
-    private let store: ImageDataStore
-    
-    init(store: ImageDataStore) {
-        self.store = store
-    }
-    
-    func loadImageData(from url: URL, completion: @escaping (ImageLoader.Result) -> Void) -> ImageLoaderTask {
-        let task = Task(completion)
-        store.retrieve(dataForURL: url) { [weak self] result in
-            guard self != nil else { return }
-            task.complete(with: result
-                            .mapError { _ in Error.failed }
-                            .flatMap { data in
-                data.map { .success($0) } ?? .failure(Error.notFound)
-            })
-        }
-        return task
-    }
-}
-
 class LocalImageLoaderTests: XCTestCase {
     
     func test_init_doesNotMessageStoreUponCreation() {
